Merge branch 'develop' into pnpm ### Issue / feature description
`Quick description of the issue or the feature that is implemented in the pr ie: add new transitions between screens`

### Changes

`quick list regarding the related changes within the pr`

ie:

- add new transaction transformer
- add new ui

### Checklist

- [ ] Rebased to the last commit of the target branch (or merged)
- [ ] Code self-reviewed
- [ ] Code self-tested
- [ ] Tests updated (if needed)
- [ ] All tests are passing locally

Please keep your pull request as small as possible. If you need to make multiple changes, please create separate pull requests for each change. Create a draft pull request if you need early feedback. This will mark the pull request as a work in progress and prevent it from being reviewed or merged until you are ready.

Please move drafts to the ready for review (regular PR) when you are ready to start the review process and other developers will pick it up from there.